refactor main code
