+++ src/main/java/duke/Duke.java
@@ -4,6 +4,12 @@ import java.util.ListIterator;
 
 import java.nio.file.Paths;
 
+import task.Task;
+import utility.Parser;
+import utility.Storage;
+import utility.TaskList;
+import utility.Ui;
+
 /**
  * Represents the highest level code responsible for Duke's operations.
  */
