+++ src/main/java/duke/MainWindow.java
@@ -8,6 +8,8 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
+import utility.Ui;
+
 /**
  * Controller for MainWindow. Provides the layout for the other controls.
  */
