+++ src/main/java/task/Event.java
@@ -1,4 +1,4 @@
-package duke;
+package task;
 
 import java.time.LocalDate;
