+++ src/main/java/utility/Parser.java
@@ -1,9 +1,17 @@
-package duke;
+package utility;
 
 import java.time.LocalDate;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+
+import duke.Command;
+import task.Deadline;
+import duke.DukeCommand;
+import duke.DukeException;
+import task.Event;
+import task.Task;
+import task.Todo;
 
 /**
  * Parser is involved with parsing user commands in Duke.
