+++ src/main/java/utility/Storage.java
@@ -1,4 +1,4 @@
-package duke;
+package utility;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
@@ -9,6 +9,11 @@ import java.nio.file.Path;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
+
+import task.Deadline;
+import task.Event;
+import task.Task;
+import task.Todo;
 
 /**
  * Handles read and write operations for Task objects in Duke.
