+++ src/test/java/duke/DukeExceptionTest.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.params.provider.ValueSource;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import utility.Parser;
+
 public class DukeExceptionTest {
 
     @ParameterizedTest
